Make less assumptions about Steam load times
Simply waiting 30 seconds and expecting the cache folder to be created is unrealistic for slower systems and internet connections, or in the case that Steam is having connectivity issues, so let's wait for the cache folder to be recreated and then wait an additional 5 seconds just to ensure that the friends.css has been cached. Needs more work, but this method should be better than the last.