Begin modularization process
We probably shouldn't be doing everything in main, for now let's split up the finding and patching function into its own method so that we don't have to restart the entire program to redetect and patch the cache file.

Ideally both locating the file and patching it would be in separate functions.

Currently we're redownloading the original CSS file from Steam every time PatchCacheFile() is executed on the off chance that the CSS file changes during Steam's restart, and this also helps pad out time in case the cache isn't ready.