Add basic friends window check & attempt to automatically reopen friends window
One potential issue for this function is that if the user gets a message sent to them between the friends list opening and the program detecting this, it will fail to detect the window as the window title will change.

Checking window titles with a regular expression is a lot more involved, and this is a fairly unlikely scenario due to how fast it should detect the window, so I think this is fine for now.