Safer find window method
Thanks to FindWindowLike class from Bob Learned on Experts Exchange for the find window code. Now it will find the friends list even if a chat is opened and the window title is different.

Now it will also keep opened chats open when it reopens the window, at the cost of a slightly longer reopen time (since it now has to reconnect to Steam)