Attempt to fix CSS corruption
Recompressing the CSS file resulted in certain visual elements breaking (Notification Settings, Multi User Chats). For now, the solution will be to import an untouched version of the original, but this is less than ideal as it results in the program constantly finding more matches on future runs even if no update is required.

Also change compression to close file stream afterwards, and use FS to write the file to disk. This should be more efficient and safer.