Fixed invalid URL
Steam-Chat is being used to confirm the friends.css used by the client is the one downloaded by the patcher for comparison. In the past it worked without any arguments, however now without them it returns an error page that does not include a reference to the friends.css.

Let's include arguments so that we can get the correct CSS.