Better fix for compression issues
Our compression method is faulty. The previous fix worked for the settings window but left multi user chats broken. To fix this, let's copy the friends.css to clientui/friends.original.css and load it from there in the cache file, 'commenting' out the original css in there with unlabeled brackets ( { } ) so that it is not loaded.

This way, steamloopback.host can handle compression of the css and we don't have to worry about it. Alternatively, we could load the CSS directly from Steam's CDN, however if we do this it will always be detected upon rerunning the patcher because it will always match.

In the original.friends.css, we write the etag on the top of the file as a comment to prevent it from matching.

Thoughts and suggestions on this welcome.